Cutover to new TemporalMemory by deleting old TM and Tests
